Make page API specs via MS

+++ Bikewale.Cache/BikeData/BikeMakesCacheRepository.cs
@@ -79,13 +79,13 @@ namespace Bikewale.Cache.BikeData
         /// </summary>
         /// <param name="makeId"></param>
         /// <returns></returns>
-        public BikeDescriptionEntity GetMakeDescription(uint makeId)
+        public BikeDescriptionEntity GetMakeDescription(int makeId)
         {
             BikeDescriptionEntity objMakeDesc = null;
             string key = String.Format("BW_MakeDescription_{0}", makeId);
             try
             {
-                objMakeDesc = _cache.GetFromCache<BikeDescriptionEntity>(key, new TimeSpan(1, 0, 0), () => _objMakes.GetMakeDescription((int)makeId));
+                objMakeDesc = _cache.GetFromCache<BikeDescriptionEntity>(key, new TimeSpan(1, 0, 0), () => _objMakes.GetMakeDescription(makeId));
             }
             catch (Exception ex)
             {
